Wack Wack Wack. More and more tests passes now :)

+++ pypy/translator/js/test/test_rclass.py
@@ -2,8 +2,6 @@ import py
 from pypy.translator.js.test.runtest import JsTest
 from pypy.rpython.test.test_exception import BaseTestException
 from pypy.rpython.test.test_rclass import BaseTestRclass
-from pypy.rpython.test.test_rlist import BaseTestRlist
-from pypy.rpython.test.test_rpbc import BaseTestRPBC
 from pypy.rpython.test.test_rtuple import BaseTestRtuple
 from pypy.rpython.test.test_rstr import BaseTestRstr
 
@@ -64,7 +62,7 @@ class TestJsClass(JsTest, BaseTestRclass):
             return v0, v1, v2
 
         res = self.interpret(f, [])
-        assert isinstance(res[0], float)
+        assert isinstance(res[0], int)
 
     def test_hash_preservation(self):
         py.test.skip("WIP")
@@ -75,10 +73,6 @@ class TestJsClass(JsTest, BaseTestRclass):
     def test_isinstance(self):
         py.test.skip("WIP")
 
-#class TestJsList(JsTest, BaseTestRlist):
-#    def test_insert_bug(self):
-#        py.test.skip("in progress")
-##    
 #class TestJsPBC(JsTest, BaseTestRPBC):
 #    pass
 ##
